+++ supportedBackends/aws/src/main/scala/cromwell/backend/impl/aws/AwsBatchAsyncBackendJobExecutionActor.scala
@@ -157,6 +157,7 @@ class AwsBatchAsyncBackendJobExecutionActor(
   override lazy val dockerImageUsed: Option[String] = Option(jobDockerImage)
 
   // overriden function : used in StandardAsyncExecutionActor
+  // AWS localization optional came in with the fork and is untested as of 2025-11 (no Centaur cases)
   override def inputsToNotLocalize: Set[WomFile] =
     jobDescriptor.fullyQualifiedInputs.collect {
       case (_, womFile: WomFile)
@@ -260,7 +261,7 @@ class AwsBatchAsyncBackendJobExecutionActor(
       configuration.efsMntPoint,
       Option(runtimeAttributes.efsMakeMD5),
       Option(runtimeAttributes.efsDelocalize),
-      Option(runtimeAttributes.tagResources),
+      configuration.tagResources,
       runtimeAttributes.logGroupName,
       runtimeAttributes.additionalTags,
       scriptBucketPrefix