+++ supportedBackends/aws/src/test/scala/cromwell/backend/impl/aws/AwsBatchAttributesSpec.scala
@@ -53,6 +53,7 @@ class AwsBatchAttributesSpec extends AnyFlatSpec with CromwellTimeoutSpec with M
 
     val attributes = AwsBatchAttributes.fromConfigs(config, backendConfig)
     attributes.executionBucket should be("s3://myBucket")
+    attributes.tagResources should be(Some(true))
   }
 
   it should "not parse invalid config" taggedAs IntegrationTest in {
@@ -76,6 +77,7 @@ class AwsBatchAttributesSpec extends AnyFlatSpec with CromwellTimeoutSpec with M
        |   maximum-polling-interval = 600
        |   numSubmitAttempts = 6
        |   numCreateDefinitionAttempts = 6
+       |   tagResources = true
        |
        |
        |   filesystems = {